requirements.txt
pip~=23.2.1
wheel~=0.41.2
setuptools~=68.2.0
pillow~=10.3.0
pyqt6